Removed underscores from all component names
